added patch function
